+++ frontend/public/changelog.md
@@ -9,6 +9,7 @@
 - Memberikan tautan ke KBBI di detail kamus
 - Menambahkan kolom sumber untuk entri kamus dengan standar KBBI IV
 - Menambahkan jenis entri prakategorial dan menambahkannya ke kategori "Bentuk Terikat"
+- Menambahkan filter jenis di redaksi kamus
 
 ## 2026-02-18 Rabu